+++ CcApps/CcSync/CcSync/src/CcSyncServerWorker.cpp
@@ -225,7 +225,7 @@ bool CcSyncServerWorker::receiveFile(CcFile* pFile, CcSyncFileInfo& oFileInfo)
   bool bRet = false;
   bool bTransfer = true;
   CcCrc32 oCrc;
-  size_t uiReceived = 0;
+  uint64 uiReceived = 0;
   while (bTransfer)
   {
     if (uiReceived < oFileInfo.getSize())
@@ -233,7 +233,7 @@ bool CcSyncServerWorker::receiveFile(CcFile* pFile, CcSyncFileInfo& oFileInfo)
       size_t uiBufferSize = CcSyncGlobals::TransferSize;
       if (oFileInfo.getSize() - uiReceived < CcSyncGlobals::TransferSize)
       {
-        uiBufferSize = oFileInfo.getSize() - uiReceived;
+        uiBufferSize = static_cast<size_t>(oFileInfo.getSize() - uiReceived);
       }
       CcByteArray oByteArray(uiBufferSize);
       m_oSocket->readArray(oByteArray);
